Refactor [form schema] assignment_schema form
