+++ pages/edu/assignment_schema/edit/[spec].tsx
@@ -1,6 +1,6 @@
 import { useLocale } from '@hooks/useLocale';
 import { ReactNode, useCallback, useEffect, useMemo } from 'react';
-import { useForm } from '@mantine/form';
+import { useForm, UseFormReturnType } from '@mantine/form';
 import { ITaskDisplay } from '@custom-types/data/ITask';
 import { DefaultLayout } from '@layouts/DefaultLayout';
 import { Item } from '@ui/CustomTransferList/CustomTransferList';
@@ -39,62 +39,40 @@ function EditAssignmentSchema({
     }),
     [assignment_schema]
   );
-  const form = useForm({
-    initialValues: formValues,
-    validate: {
-      title: (value) =>
-        value.length < 5
-          ? locale.assignmentSchema.form.validation.title
-          : null,
-      description: (value) =>
-        value.length < 20
-          ? locale.assignmentSchema.form.validation.description
-          : null,
-      tasks: (value) =>
-        value.length === 0
-          ? locale.assignmentSchema.form.validation.tasks
-          : null,
-      defaultDuration: (value) =>
-        value <= 5
-          ? locale.assignmentSchema.form.validation.defaultDuration
-          : null,
+
+  const handleSubmit = useCallback(
+    (form: UseFormReturnType<any>) => {
+      if (form.validate().hasErrors) {
+        const id = newNotification({});
+        errorNotification({
+          id,
+          title: locale.validationError,
+          autoClose: 5000,
+        });
+        return;
+      }
+      let body: any = {
+        ...form.values,
+        tasks: form.values['tasks'].map((task: Item) => task.value),
+        tags: form.values['tags'].map((tag: Item) => tag.value),
+        defaultDuration: form.values.defaultDuration * 60 * 1000, // from minutes to milliseconds
+      };
+      requestWithNotify(
+        `assignment_schema/edit`,
+        'POST',
+        locale.notify.assignmentSchema.edit,
+        lang,
+        (response: IAssignmentSchema) => response.spec,
+        body
+      );
     },
-  });
-
-  useEffect(() => {
-    form.setValues(formValues);
-  }, [formValues]); // eslint-disable-line
-
-  const handleSubmit = useCallback(() => {
-    if (form.validate().hasErrors) {
-      const id = newNotification({});
-      errorNotification({
-        id,
-        title: locale.validationError,
-        autoClose: 5000,
-      });
-      return;
-    }
-    let body: any = {
-      ...form.values,
-      tasks: form.values['tasks'].map((task: Item) => task.value),
-      tags: form.values['tags'].map((tag: Item) => tag.value),
-      defaultDuration: form.values.defaultDuration * 60 * 1000, // from minutes to milliseconds
-    };
-    requestWithNotify(
-      `assignment_schema/edit`,
-      'POST',
-      locale.notify.assignmentSchema.edit,
-      lang,
-      (response: IAssignmentSchema) => response.spec,
-      body
-    );
-  }, [form, locale, lang]);
+    [locale, lang]
+  );
 
   return (
     <>
       <Form
-        form={form}
+        initialValues={formValues}
         handleSubmit={handleSubmit}
         buttonLabel={locale.form.update}
       />
